trying to fix multiplayer
the problem: clients couldn't join the host's session. now they can, i didn't really change anything

next problem: clients can't actually control their character. what i think the issue was is that the player controller was executing a "run as server" function to execute a game mode function to spawn a character, and when passing the necessary information to the second function, it used a self reference for player controller, which was assigned to the new pawn by the game mode. however, the self reference was inside the "run as server" function, so it's entirely likely that the controller assigned to the new pawn was the server-side version of the controller, which the player has no access to, instead of the client-side player controller. i now have the self reference being passed to the "run as server" function and then on to the game mode function, so the possess node should tie back to the correct controller

there's still an in-editor lag for picking up shares that does not appear to exist in the packaged version, and some rubberbanding for no obvious reason

also, everything about the game was laid out to fill exactly a 1920x1080 resolution, and will not scale properly up or down, which is only a problem if your monitor is smaller than that (like one of mine is)

also just for fun i tried to simplify spawning the characters by using a select statement to return the character class, instead of a switch statement to spawn joe and possess or louise and possess, which is twice as many nodes as needed. however, when you don't explicitly specify the actor to be spawned, the output can't be piped straight to the possess node because it isn't compatible. i wonder if i could drill down into the output and get the part i need, though